FEATURE: Allow the current timestamp to be injected
Extends the constructor of the `DoctrineEventStore` by
a parameter `$clock` that can be set to an instance of
the `ClockInterface` (https://www.php-fig.org/psr/psr-20/).

This allows the `recordedAt` field of events to be tampered
with (which is especially useful while testing).

The parameter has a default value. So the only "breaking" change
is the new composer dependency to `psr/clock`.